requirements.txt
absl-py==0.9.0
alembic==1.4.2
astor==0.8.1
attrdict==2.0.1
audioread==2.1.8
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cliff==3.2.0
cmaes==0.5.0
cmd2==0.8.9
colorlog==4.1.0
decorator==4.4.2
ds-ctcdecoder==0.7.1
gast==0.2.2
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.1
joblib==0.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
llvmlite==0.31.0
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.19.4
opt-einsum==3.2.1
optuna==1.5.0
opuslib==2.0.0
pbr==5.4.5
prettytable==0.7.2
progressbar2==3.51.3
protobuf==3.12.2
pycparser==2.20
pyperclip==1.8.0
python-dateutil==2.8.1
python-editor==1.0.4
python-utils==2.4.0
pyxdg==0.26
PyYAML==5.3.1
requests==2.23.0
resampy==0.2.2
scikit-learn==0.23.1
semver==2.10.1
six==1.15.0
SoundFile==0.10.3.post1
soupsieve==2.0.1
sox==1.3.7
SQLAlchemy==1.3.17
stevedore==2.0.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.2
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.51.0
urllib3==1.25.9
wcwidth==0.2.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
librosa==0.8.0